listing func refac, wip
