+++ testutils/testing_utils.go
@@ -12,6 +12,7 @@ import (
 	"github.com/eris-ltd/eris-cli/config"
 	def "github.com/eris-ltd/eris-cli/definitions"
 	ini "github.com/eris-ltd/eris-cli/initialize"
+	"github.com/eris-ltd/eris-cli/list"
 	"github.com/eris-ltd/eris-cli/util"
 
 	log "github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/Sirupsen/logrus"
@@ -217,7 +218,7 @@ func checkIPFSnotRunning() {
 	do.Quiet = true
 	do.Operations.Args = []string{"testing"}
 	log.Debug("Finding the running services")
-	if err := util.ListAll(do, "services"); err != nil {
+	if err := list.ListAll(do, "services"); err != nil {
 		IfExit(err)
 	}
 	res := strings.Split(do.Result, "\n")
@@ -234,7 +235,7 @@ func checkIPFSnotRunning() {
 	do.Quiet = true
 	do.Operations.Args = []string{"testing"}
 	log.Debug("Finding the existing services")
-	if err := util.ListAll(do, "services"); err != nil {
+	if err := list.ListAll(do, "services"); err != nil {
 		IfExit(err)
 	}
 	res = strings.Split(do.Result, "\n")
