+++ util/container_stats.go
@@ -1,7 +1,7 @@
 package util
 
 import (
-	"bytes"
+	//"bytes"
 	"fmt"
 	"regexp"
 	"sort"
@@ -14,7 +14,7 @@ import (
 	log "github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/Sirupsen/logrus"
 	"github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/fsouza/go-dockerclient"
 	"github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/oleiade/reflections"
-	"github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/olekukonko/tablewriter"
+	//"github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/olekukonko/tablewriter"
 	"github.com/eris-ltd/eris-cli/Godeps/_workspace/src/github.com/serenize/snaker"
 )
 
@@ -53,47 +53,6 @@ func PrintInspectionReport(cont *docker.Container, field string) error {
 	}
 
 	return nil
-}
-
-func PrintTableReport(typ string, existing, all bool) (string, error) {
-	log.WithField("type", typ).Debug("Table report initialized")
-
-	var conts []*ContainerName
-	if !all {
-		conts = ErisContainersByType(typ, existing)
-	}
-
-	buf := new(bytes.Buffer)
-	table := tablewriter.NewWriter(buf)
-	table.SetHeader([]string{fmt.Sprintf("%s NAME", strings.ToUpper(typ)), "TYPE", "RUNNING", "CONTAINER NAME", "CONTAINER #", "PORTS"})
-
-	if all { //get all the things
-		parts, _ := AssembleTable(typ)
-		for _, p := range parts {
-			table.Append(formatLine(p))
-		}
-	} else {
-		for _, c := range conts {
-			n, _ := PrintLineByContainerName(c.FullName, existing)
-			if typ == "chain" {
-				head, _ := GetHead()
-				if n[0] == head {
-					n[0] = fmt.Sprintf("**  %s", n[0]) // TODO: colorize this when we settle on a lib
-				}
-			}
-			table.Append(n)
-		}
-	}
-
-	// Styling
-	table.SetBorder(false)
-	table.SetCenterSeparator(" ")
-	table.SetColumnSeparator(" ")
-	table.SetRowSeparator("-")
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
-	table.Render()
-
-	return buf.String(), nil
 }
 
 func PrintLineByContainerName(containerName string, existing bool) ([]string, error) {
@@ -176,7 +135,7 @@ func printLine(container *docker.Container, existing bool) ([]string, error) {
 
 	Names := ContainerDisassemble(n)
 
-	parts := []string{Names.ShortName, Names.Type, running, Names.FullName, fmt.Sprintf("%d", Names.Number), formulatePortsOutput(container)}
+	parts := []string{Names.ShortName, Names.Type, running, Names.FullName, fmt.Sprintf("%d", Names.Number), FormulatePortsOutput(container)}
 	return parts, nil
 }
 
@@ -240,7 +199,7 @@ func probablyHasDataContainer(container *docker.Container) bool {
 	return false
 }
 
-func formulatePortsOutput(container *docker.Container) string {
+func FormulatePortsOutput(container *docker.Container) string {
 	var ports string
 	for k, v := range container.NetworkSettings.Ports {
 		if len(v) != 0 {
@@ -331,116 +290,4 @@ func listContainers(all bool) []docker.APIContainers {
 	}
 
 	return container
-}
-
-//----------------------------------------------------------
-//---------------------helpers for ls w/o flags-------------
-
-//XXX test with multiple containers of same definition!
-func AssembleTable(typ string) ([]Parts, error) {
-	// []string
-	known := GetGlobalLevelConfigFilesByType(typ, false) // definition files
-
-	typ = strings.TrimSuffix(typ, "s") // :(
-	// []*ContainerName
-	contsR := ErisContainersByType(typ, false) //running
-	contsE := ErisContainersByType(typ, true)  //existing
-
-	if len(contsE) == 0 && len(contsR) == 0 && len(known) == 0 {
-		return []Parts{}, nil
-	}
-
-	var myTable []Parts
-	addedAlready := make(map[string]bool)
-
-	if typ != "data" {
-		for _, name := range contsR {
-			part, _ := makePartFromContainer(name.FullName, false)
-			addedAlready[part.ShortName] = true //has to come after because short name needed
-			part.Running = true
-			myTable = append(myTable, part)
-		}
-	}
-
-	for _, name := range contsE {
-		part, _ := makePartFromContainer(name.FullName, false)
-		if addedAlready[part.ShortName] == true {
-			continue
-		} else {
-			part.Running = false
-			myTable = append(myTable, part)
-		}
-	}
-
-	if typ != "data" {
-		for _, name := range known {
-			if addedAlready[name] == true { //name from known == part.ShortName
-				continue
-			}
-			part, _ := makePartFromContainer(name, true)
-			part.Running = false
-			myTable = append(myTable, part)
-		}
-	}
-
-	return myTable, nil
-}
-
-func formatLine(p Parts) []string {
-	var running string
-	if p.Running {
-		running = "Yes"
-	} else {
-		running = "No"
-	}
-	number := fmt.Sprintf("%d", p.Number)
-
-	//must match header
-	part := []string{p.ShortName, p.Type, running, p.FullName, number, p.PortsOutput}
-
-	if p.Type == "chain" {
-		head, _ := GetHead()
-		if part[0] == head {
-			part[0] = fmt.Sprintf("**  %s", part[0]) // TODO: colorize this when we settle on a lib
-		}
-	}
-	return part
-}
-
-func makePartFromContainer(name string, defs bool) (v Parts, err error) {
-	if defs {
-		v.ShortName = name // the only value needed for known
-	} else {
-		// this block pulls out functionality from
-		// PrintLineByContainerName{Id} & printLine
-		var contID *docker.Container
-		cont, exists := ParseContainers(name, true)
-		if exists {
-			contID, err = DockerClient.InspectContainer(cont.ID)
-			if err != nil {
-				return Parts{}, err
-			}
-		}
-		if err != nil {
-			return Parts{}, err
-		}
-		tmp, err := reflections.GetField(contID, "Name")
-		if err != nil {
-			return Parts{}, err
-		}
-
-		n := tmp.(string)
-
-		Names := ContainerDisassemble(n)
-
-		v = Parts{
-			ShortName: Names.ShortName,
-			Type:      Names.Type,
-			//Running: set in previous function
-			FullName:    Names.FullName,
-			Number:      Names.Number,
-			PortsOutput: formulatePortsOutput(contID),
-		}
-	}
-	return v, nil
 }
